+++ mrminio/file_provider.go
@@ -4,9 +4,10 @@ import (
 	"context"
 	"fmt"
 	"path/filepath"
+	"strings"
 
 	"github.com/minio/minio-go/v7"
-	"github.com/mondegor/go-webcore/mrcore"
+	"github.com/mondegor/go-storage/mrstorage"
 	"github.com/mondegor/go-webcore/mrlib"
 	"github.com/mondegor/go-webcore/mrtype"
 )
@@ -14,64 +15,84 @@ import (
 // https://min.io/docs/minio/linux/developers/go/API.html
 
 type (
-	fileProvider struct {
+	FileProvider struct {
 		*ConnAdapter
 		bucketName string
+		baseDir    string
 	}
 )
 
-func NewFileProvider(conn *ConnAdapter, bucketName string) *fileProvider {
-	return &fileProvider{
+func NewFileProvider(conn *ConnAdapter, bucketName string) *FileProvider {
+	return &FileProvider{
 		ConnAdapter: conn,
 		bucketName:  bucketName,
 	}
 }
 
-func (fp *fileProvider) Info(ctx context.Context, path string) (mrtype.FileInfo, error) {
+func (fp *FileProvider) WithBaseDir(value string) (mrstorage.ExtFileProviderAPI, error) {
+	if value != "" {
+		value = strings.Trim(value, "/")
+
+		if value != "" {
+			value += "/"
+		}
+	}
+
+	if fp.baseDir == value {
+		return fp, nil
+	}
+
+	c := *fp
+	c.baseDir = value
+
+	return &c, nil
+}
+
+func (fp *FileProvider) Info(ctx context.Context, fileName string) (mrtype.FileInfo, error) {
 	info, err := fp.conn.StatObject(
 		ctx,
 		fp.bucketName,
-		path,
+		fp.baseDir+fileName,
 		minio.StatObjectOptions{},
 	)
 
 	if err != nil {
-		return mrtype.FileInfo{}, err
+		return mrtype.FileInfo{}, fp.wrapError(err)
 	}
 
-	return fp.getFileInfo(&info, path), nil
+	return fp.getFileInfo(&info, fileName), nil
 }
 
-func (fp *fileProvider) Download(ctx context.Context, path string) (*mrtype.File, error) {
+func (fp *FileProvider) Download(ctx context.Context, fileName string) (*mrtype.File, error) {
 	object, err := fp.conn.GetObject(
 		ctx,
 		fp.bucketName,
-		path,
+		fp.baseDir+fileName,
 		minio.GetObjectOptions{},
 	)
 
 	if err != nil {
-		return nil, mrcore.FactoryErrInternal.Wrap(err) // :TODO: обернуть ошибку объект не найден
+		return nil, fp.wrapError(err)
 	}
 
 	info, err := object.Stat()
 
 	if err != nil {
 		object.Close()
-		return nil, err
+		return nil, fp.wrapError(err)
 	}
 
 	return &mrtype.File{
-		FileInfo: fp.getFileInfo(&info, path),
+		FileInfo: fp.getFileInfo(&info, fileName),
 		Body:     object,
 	}, nil
 }
 
-func (fp *fileProvider) Upload(ctx context.Context, file *mrtype.File) error {
+func (fp *FileProvider) Upload(ctx context.Context, file *mrtype.File) error {
 	_, err := fp.conn.PutObject(
 		ctx,
 		fp.bucketName,
-		file.Path,
+		fp.baseDir+file.Path,
 		file.Body,
 		file.Size, // -1 - calculating size
 		minio.PutObjectOptions{
@@ -80,21 +101,31 @@ func (fp *fileProvider) Upload(ctx context.Context, file *mrtype.File) error {
 		},
 	)
 
-	return err
+	if err != nil {
+		return fp.wrapError(err)
+	}
+
+	return nil
 }
 
-func (fp *fileProvider) Remove(ctx context.Context, path string) error {
-	return fp.conn.RemoveObject(
+func (fp *FileProvider) Remove(ctx context.Context, fileName string) error {
+	err := fp.conn.RemoveObject(
 		ctx,
 		fp.bucketName,
-		path,
+		fp.baseDir+fileName,
 		minio.RemoveObjectOptions{},
 	)
+
+	if err != nil {
+		return fp.wrapError(err)
+	}
+
+	return nil
 }
 
-func (fp *fileProvider) getFileInfo(info *minio.ObjectInfo, path string) mrtype.FileInfo {
+func (fp *FileProvider) getFileInfo(info *minio.ObjectInfo, fileName string) mrtype.FileInfo {
 	return mrtype.FileInfo{
-		ContentType:  fp.getContentType(info.ContentType, path),
+		ContentType:  fp.getContentType(info.ContentType, fileName),
 		OriginalName: fp.getOriginalName(info.Metadata.Get("Content-Disposition")),
 		Name:         info.Key,
 		LastModified: info.LastModified,
@@ -102,15 +133,15 @@ func (fp *fileProvider) getFileInfo(info *minio.ObjectInfo, path string) mrtype.
 	}
 }
 
-func (fp *fileProvider) getContentType(value, path string) string {
+func (fp *FileProvider) getContentType(value, fileName string) string {
 	if value != "" {
 		return value
 	}
 
-	return mrlib.MimeTypeByExt(filepath.Ext(path))
+	return mrlib.MimeTypeByExt(filepath.Ext(fileName))
 }
 
-func (fp *fileProvider) getContentDisposition(value string) string {
+func (fp *FileProvider) getContentDisposition(value string) string {
 	if value == "" {
 		return ""
 	}
@@ -118,7 +149,7 @@ func (fp *fileProvider) getContentDisposition(value string) string {
 	return fmt.Sprintf("attachment; filename=\"%s\"", value) // :TODO: escape value
 }
 
-func (fp *fileProvider) getOriginalName(contentDisposition string) string {
+func (fp *FileProvider) getOriginalName(contentDisposition string) string {
 	const prefix = "attachment; filename=\""
 	const minLength = 23 // len of prefix + '"'